+++ src/pymetropolis/metro_simulation/demand/trips.py
@@ -23,7 +23,10 @@ from pymetropolis.metro_demand.modes.files import (
     WalkingTravelTimesFile,
 )
 from pymetropolis.metro_demand.population import TripsFile
-from pymetropolis.metro_demand.routing.files import TripsRoadNodesFile
+from pymetropolis.metro_demand.routing.files import (
+    NonPrimaryCarTrips,
+    PrimaryCarTripsAccessEgressFile,
+)
 from pymetropolis.metro_pipeline.file import MetroDataFrameFile
 from pymetropolis.metro_pipeline.steps import InputFile
 from pymetropolis.metro_simulation.common import StepWithModes, StepWithRidesharingCount
@@ -39,7 +42,8 @@ def generate_car_trips(
     mode: str,
     vehicle_type: str,
     df: pl.DataFrame,
-    ods_file: TripsRoadNodesFile,
+    primary_trips_file: PrimaryCarTripsAccessEgressFile,
+    secondary_trips_file: NonPrimaryCarTrips,
     pref_file: MetroDataFrameFile,
     tstars_file: TstarsFile,
     schedule_pref_file: LinearScheduleFile,
@@ -48,22 +52,49 @@ def generate_car_trips(
 ):
     import polars as pl
 
-    df = df.with_columns(
-        pl.lit(mode).alias("alt_id"),
-        pl.lit("Road").alias("class.type"),
+    df = df.with_columns(pl.lit(mode).alias("alt_id"))
+    primary_trips: pl.DataFrame = primary_trips_file.read().select(
+        "trip_id",
+        pl.col("access_node").cast(pl.String).alias("class.origin"),
+        pl.col("egress_node").cast(pl.String).alias("class.destination"),
+        "access_time",
+        "egress_time",
         pl.lit(vehicle_type).alias("class.vehicle"),
     )
-    ods: pl.DataFrame = ods_file.read()
-    df = df.join(ods, on="trip_id", how="left").rename(
-        {"origin_road_node": "class.origin", "destination_road_node": "class.destination"}
+    df = df.join(primary_trips, on="trip_id", how="left")
+    secondary_trips: pl.DataFrame = secondary_trips_file.read().select(
+        "trip_id", pl_duration_to_seconds("free_flow_travel_time").alias("class.travel_time")
+    )
+    df = df.join(secondary_trips, on="trip_id", how="left")
+    df = df.with_columns(
+        pl.when(pl.col("class.origin").is_not_null())
+        .then(pl.lit("Road"))
+        .when(pl.col("class.travel_time").is_not_null())
+        .then(pl.lit("Virtual"))
+        .alias("class.type"),
+        access_time_sec=pl_duration_to_seconds("access_time").fill_null(0.0),
+        egress_time_sec=pl_duration_to_seconds("egress_time").fill_null(0.0),
+    ).drop("access_time", "egress_time")
+    # Drop trips which are neither primary nor secondary.
+    df = df.filter(pl.col("class.type").is_not_null())
+    # Compute stopping time at destination: egress time + activity time + next access time.
+    df = df.with_columns(
+        stopping_time=pl.col("egress_time_sec")
+        + pl.col("activity_time")
+        + pl.col("access_time_sec").shift(-1).over("agent_id").fill_null(0.0)
     )
     if pref_file.exists():
         params: pl.DataFrame = pref_file.read().select(
             "person_id",
-            constant_utility=-pl.col(f"{mode}_cst"),
+            constant_utility=pl.col(f"{mode}_cst"),
             alpha=pl.col(f"{mode}_vot") / 3600.0,
         )
         df = df.join(params, on="person_id", how="left")
+        # Decrease utility by the access and egress time's value of time.
+        df = df.with_columns(
+            constant_utility=pl.col("constant_utility")
+            - pl.col("alpha") * (pl.col("access_time_sec") + pl.col("egress_time_sec"))
+        )
     if tstars_file.exists():
         tstars: pl.DataFrame = tstars_file.read()
         df = (
@@ -80,7 +111,7 @@ def generate_car_trips(
                 pl.lit("Linear").alias("schedule_utility.type"),
                 (pl.col("beta") / 3600.0).alias("schedule_utility.beta"),
                 (pl.col("gamma") / 3600.0).alias("schedule_utility.gamma"),
-                pl.col("delta").dt.total_seconds().cast(pl.Float64).alias("schedule_utility.delta"),
+                pl_duration_to_seconds("delta").alias("schedule_utility.delta"),
             )
             .drop("beta", "gamma", "delta")
         )
@@ -98,6 +129,7 @@ def generate_car_trips(
             )
             .drop("fuel_cost")
         )
+    df = df.drop("activity_time", "access_time_sec", "egress_time_sec")
     return df
 
 
@@ -113,15 +145,12 @@ def generate_public_transit_trips(
 
     df = df.with_columns(
         pl.lit("public_transit").alias("alt_id"), pl.lit("Virtual").alias("class.type")
-    )
+    ).rename({"activity_time": "stopping_time"})
     tts: pl.DataFrame = tts_file.read()
     df = (
         df.join(tts, on="trip_id", how="left")
         .with_columns(
-            pl.col("public_transit_travel_time")
-            .dt.total_seconds()
-            .cast(pl.Float64)
-            .alias("class.travel_time")
+            pl_duration_to_seconds("public_transit_travel_time").alias("class.travel_time")
         )
         .drop("public_transit_travel_time")
     )
@@ -150,7 +179,7 @@ def generate_public_transit_trips(
                 pl.lit("Linear").alias("schedule_utility.type"),
                 (pl.col("beta") / 3600.0).alias("schedule_utility.beta"),
                 (pl.col("gamma") / 3600.0).alias("schedule_utility.gamma"),
-                pl.col("delta").dt.total_seconds().cast(pl.Float64).alias("schedule_utility.delta"),
+                pl_duration_to_seconds("delta").alias("schedule_utility.delta"),
             )
             .drop("beta", "gamma", "delta")
         )
@@ -167,16 +196,13 @@ def generate_walking_trips(
 ):
     import polars as pl
 
-    df = df.with_columns(pl.lit("walking").alias("alt_id"), pl.lit("Virtual").alias("class.type"))
+    df = df.with_columns(
+        pl.lit("walking").alias("alt_id"), pl.lit("Virtual").alias("class.type")
+    ).rename({"activity_time": "stopping_time"})
     tts: pl.DataFrame = tts_file.read()
     df = (
         df.join(tts, on="trip_id", how="left")
-        .with_columns(
-            pl.col("walking_travel_time")
-            .dt.total_seconds()
-            .cast(pl.Float64)
-            .alias("class.travel_time")
-        )
+        .with_columns(pl_duration_to_seconds("walking_travel_time").alias("class.travel_time"))
         .drop("walking_travel_time")
     )
     if pref_file.exists():
@@ -203,7 +229,7 @@ def generate_walking_trips(
                 pl.lit("Linear").alias("schedule_utility.type"),
                 (pl.col("beta") / 3600.0).alias("schedule_utility.beta"),
                 (pl.col("gamma") / 3600.0).alias("schedule_utility.gamma"),
-                pl.col("delta").dt.total_seconds().cast(pl.Float64).alias("schedule_utility.delta"),
+                pl_duration_to_seconds("delta").alias("schedule_utility.delta"),
             )
             .drop("beta", "gamma", "delta")
         )
@@ -220,16 +246,13 @@ def generate_bicycle_trips(
 ):
     import polars as pl
 
-    df = df.with_columns(pl.lit("bicycle").alias("alt_id"), pl.lit("Virtual").alias("class.type"))
+    df = df.with_columns(
+        pl.lit("bicycle").alias("alt_id"), pl.lit("Virtual").alias("class.type")
+    ).rename({"activity_time": "stopping_time"})
     tts: pl.DataFrame = tts_file.read()
     df = (
         df.join(tts, on="trip_id", how="left")
-        .with_columns(
-            pl.col("bicycle_travel_time")
-            .dt.total_seconds()
-            .cast(pl.Float64)
-            .alias("class.travel_time")
-        )
+        .with_columns(pl_duration_to_seconds("bicycle_travel_time").alias("class.travel_time"))
         .drop("bicycle_travel_time")
     )
     if pref_file.exists():
@@ -256,7 +279,7 @@ def generate_bicycle_trips(
                 pl.lit("Linear").alias("schedule_utility.type"),
                 (pl.col("beta") / 3600.0).alias("schedule_utility.beta"),
                 (pl.col("gamma") / 3600.0).alias("schedule_utility.gamma"),
-                pl.col("delta").dt.total_seconds().cast(pl.Float64).alias("schedule_utility.delta"),
+                pl_duration_to_seconds("delta").alias("schedule_utility.delta"),
             )
             .drop("beta", "gamma", "delta")
         )
@@ -268,8 +291,13 @@ class WriteMetroTripsStep(StepWithModes, StepWithRidesharingCount):
 
     input_files = {
         "trips": TripsFile,
-        "car_driver_ods": InputFile(
-            TripsRoadNodesFile,
+        "primary_car_trips": InputFile(
+            PrimaryCarTripsAccessEgressFile,
+            when=lambda inst: inst.has_car_mode(),
+            when_doc='if any "car_*" mode is defined',
+        ),
+        "secondary_car_trips": InputFile(
+            NonPrimaryCarTrips,
             when=lambda inst: inst.has_car_mode(),
             when_doc='if any "car_*" mode is defined',
         ),
@@ -351,7 +379,22 @@ class WriteMetroTripsStep(StepWithModes, StepWithRidesharingCount):
         import polars as pl
 
         trips: pl.DataFrame = self.input["trips"].read()
-        df = trips.select("trip_id", "person_id", agent_id="tour_id").sort("agent_id", "trip_id")
+        if "destination_activity_duration" not in trips.columns:
+            trips = trips.with_columns(
+                destination_activity_duration=pl.lit(None, dtype=pl.Duration)
+            )
+        df = trips.select(
+            "trip_id",
+            "person_id",
+            agent_id="tour_id",
+            activity_time=pl_duration_to_seconds("destination_activity_duration").fill_null(0.0),
+        ).sort("agent_id", "trip_id")
+        # Set activity time to 0 for the last trip of the tour.
+        df = df.with_columns(
+            activity_time=pl.when(pl.col("trip_id") != pl.col("trip_id").last().over("agent_id"))
+            .then("activity_time")
+            .otherwise(0.0)
+        )
         metro_trips = pl.DataFrame()
         for car_mode, vehicle_type in (
             ("car_driver", "car_driver_alone"),
@@ -365,7 +408,8 @@ class WriteMetroTripsStep(StepWithModes, StepWithRidesharingCount):
                     car_mode,
                     vehicle_type,
                     df=df,
-                    ods_file=self.input["car_driver_ods"],
+                    primary_trips_file=self.input["primary_car_trips"],
+                    secondary_trips_file=self.input["secondary_car_trips"],
                     pref_file=self.input[f"{car_mode}_preferences"],
                     tstars_file=self.input["tstars"],
                     schedule_pref_file=self.input["linear_schedule"],